routing for navlinks added
