+++ src/index.js
@@ -1,3 +1,5 @@
+import { BrowserRouter as Router } from "react-router-dom";
+
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
