+++ src/screens/Testimonials.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import Navbar from "../components/Navbar";
+
+const Testimonials = () => {
+	return (
+		<div>
+			<Navbar />
+			<h1>Testimonials Page</h1>
+		</div>
+	);
+};
+
+export default Testimonials;
