+++ src/screens/Volunteers.jsx
@@ -1,5 +1,12 @@
+import Navbar from "../components/Navbar";
+
 const Volunteers = () => {
-	return <div></div>;
+	return (
+		<div>
+			<Navbar />
+			<h1>Volunteers Page</h1>
+		</div>
+	);
 };
 
 export default Volunteers;
